requirements.txt
streamlit
fastapi
uvicorn
pycryptodome
argon2-cffi
requests
pymongo